feat: move patpass comercio to new namespace
