+++ Transbank/PatpassComercio/Detail.cs
@@ -1,4 +1,4 @@
-﻿using Newtonsoft.Json;
+using Newtonsoft.Json;
 
 namespace Transbank.Patpass.Common
 {
